requirements.txt
streamlit
requests
beautifulsoup4
pandas
gspread
oauth2client
python-dotenv
xlsxwriter
